update Selenium IDE files.
