+++ tools/get-ide-command.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 
-set -eux
+set -eu
+
+base_url="https://raw.githubusercontent.com/SeleniumHQ/selenium-ide/master/packages/selenium-ide-extension/src/neo/models/Command"
 
 cd $(dirname "$0")/../src/main/resources/selenium-ide
-curl -LO https://raw.githubusercontent.com/SeleniumHQ/selenium-ide/master/packages/selenium-ide/src/neo/models/Command.js
+for f in ArgTypes.js Commands.js; do
+  ( set -x; curl -LO $base_url/$f )
+done
